Update Calculo de promedio.
Se agrego el calculo de promedio de diametro, dependiendo del trazo de lectura que se eliga.

+++ home.py
@@ -14,39 +14,37 @@ def leer_csv():
     solo_datos = []
 
     for fila in lista_datos[3:]:
-        solo_datos.append(fila)
         datos_separados = [item.split(';') for item in fila]
-        # item = datos_separados(0)
-        # print (item)
-        # distancia = datos_separados[1]
-        # print (distancia)
-        # cubierta = datos_separados[2]
-        # print (cubierta)
-        # diametro = datos_separados[3]
-        # print (diametro)
-        # overlay = datos_separados[4]    
-        # print (overlay)
-        # realdiametro = datos_separados[5]
-        # print (realdiametro)
-        # calidad = datos_separados[6]
-        # print (calidad)
-        # tipo = datos_separados[7]
-        # print (tipo)
-        # trazo = datos_separados[8]
-        # print (trazo)
-
-    print (datos_separados)
-    # Do something with the JSON data
-    titulos = solo_datos[0]
-    datos_con_titulos = []
-    for fila in solo_datos[1:]:
-        datos_con_titulos.append(dict(zip(titulos, fila)))
-        
-        # Do something with the JSON data
+        lista_separada = [element for sublist in datos_separados for item in sublist for element in item.split(',')]
+        item = lista_separada[0].replace(" ", "")
+        distancia = lista_separada[1]
+        cubierta = lista_separada[2]
+        diametro = lista_separada[3]
+        overlay = lista_separada[4]
+        realdiametro = lista_separada[5]
+        calidad = lista_separada[6]
+        tipo = lista_separada[7]
+        trazo = lista_separada[8].replace(" ", "")
+        solo_datos.append({'item': item, 'distancia': distancia, 'cubierta': cubierta, 'diametro': diametro, 'overlay': overlay, 'realdiametro': realdiametro, 'calidad': calidad, 'tipo': tipo, 'trazo': trazo})
+    # Do something with the JSON data        
     json_data = json.dumps(solo_datos)
-        
-    with open('data.json', 'w') as file:
-        json.dump(solo_datos, file)
+    if os.path.exists('./data.json'):
+        print("El archivo existe")
+    else:
+        with open('data.json', 'w') as file:
+            json.dump(solo_datos, file)
+    i = 0
+    promedio_diametro = 0
+    maximo_diametro = 0
+    for items in solo_datos:
+        if solo_datos[i]["trazo"] == "2":
+            solo_datos[i]["diametro"] = solo_datos[i]["diametro"].replace(" ", "").replace("mm", "")
+            promedio_diametro = promedio_diametro + float(solo_datos[i]["diametro"])
+            maximo_diametro = maximo_diametro + 1
+        else:
+            asd = 0
+        i = i + 1
+    print(promedio_diametro/maximo_diametro)
     return lista_datos
 
 def obtener_nombres_archivos(directorio = "./csvdata/"):
